Refactor command packages
