+++ vgscheckout/src/main/java/com/verygoodsecurity/vgscheckout/networking/command/core/VGSCheckoutCancellable.kt
@@ -1,4 +1,4 @@
-package com.verygoodsecurity.vgscheckout.networking.command
+package com.verygoodsecurity.vgscheckout.networking.command.core
 
 /**
  * A {@code VGSCancellable} is a action that can be canceled.
